+++ packages/auth/src/Auth.ts
@@ -140,23 +140,23 @@ export default class AuthClass {
         });
 
         // initiailize cognitoauth client if hosted ui options provided
-        if (oauth) {
+        // to keep backward compatibility:
+        const cognitoHostedUIConfig = oauth? (oauth['domain']? oauth : oauth.awsCognito) : undefined;
+        if (cognitoHostedUIConfig) {
             const that = this;
-
             const cognitoAuthParams = Object.assign(
                 {
                     ClientId: userPoolWebClientId,
                     UserPoolId: userPoolId,
-                    AppWebDomain: oauth.domain,
-                    TokenScopesArray: oauth.scope,
-                    RedirectUriSignIn: oauth.redirectSignIn,
-                    RedirectUriSignOut: oauth.redirectSignOut,
-                    ResponseType: oauth.responseType,
+                    AppWebDomain: cognitoHostedUIConfig['domain'],
+                    TokenScopesArray: cognitoHostedUIConfig['scope'],
+                    RedirectUriSignIn: cognitoHostedUIConfig['redirectSignIn'],
+                    RedirectUriSignOut: cognitoHostedUIConfig['redirectSignOut'],
+                    ResponseType: cognitoHostedUIConfig['responseType'],
                     Storage: this._storage
                 },
-                oauth.options
+                cognitoHostedUIConfig['options']
             );
-
             logger.debug('cognito auth params', cognitoAuthParams);
             this._cognitoAuthClient = new CognitoAuth(cognitoAuthParams);
             this._cognitoAuthClient.userhandler = {
@@ -180,6 +180,7 @@ export default class AuthClass {
                 onFailure: (err) => {
                     logger.debug("Error in cognito hosted auth response", err);
                     dispatchAuthEvent('signIn_failure', err);
+                    dispatchAuthEvent('cognitoHostedUI_failure', err);
                 }
             };
             // if not logged in, try to parse the url.
@@ -192,7 +193,12 @@ export default class AuthClass {
                     return;
                 }
                 const curUrl = window.location.href;
-                this._cognitoAuthClient.parseCognitoWebResponse(curUrl);
+                try {
+                    this._cognitoAuthClient.parseCognitoWebResponse(curUrl);
+                } catch (err) {
+                    logger.debug('something wrong when parsing the url', err);
+                    dispatchAuthEvent('parsingUrl_failure', null);
+                }
             });
         }
 
@@ -1151,16 +1157,8 @@ export default class AuthClass {
             logger.debug('no Congito User pool');
         }
         
-
-        try {
-            await Credentials.set(null, 'guest');
-        } catch (e) {
-            logger.debug('cannot load guest credentials for unauthenticated user', e);
-        } finally {
-            dispatchAuthEvent('signOut', this.user);
-            this.user = null;
-            return;
-        }
+        dispatchAuthEvent('signOut', this.user);
+        this.user = null;
     }
 
     private async cleanCachedItems() {